Move DynamoGateway into comparison/lib folder
