+++ src/comparison/lambdas/compareLambda.integration.test.ts
@@ -1,5 +1,7 @@
 /* eslint-disable jest/no-conditional-expect */
 jest.setTimeout(10000)
+import "tests/helpers/setEnvironmentVariables"
+
 import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3"
 import type { DocumentClient } from "aws-sdk/clients/dynamodb"
 import fs from "fs"
@@ -7,13 +9,12 @@ import MockDate from "mockdate"
 import lambda from "src/comparison/lambdas/compareLambda"
 import MockDynamo from "tests/helpers/MockDynamo"
 import MockS3 from "tests/helpers/MockS3"
-import "tests/helpers/setEnvironmentVariables"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
 import { ZodError } from "zod"
-import DynamoGateway from "../DynamoGateway/DynamoGateway"
 import * as CompareMessage from "../lib/compareMessage"
 import createDynamoDbConfig from "../lib/createDynamoDbConfig"
 import createS3Config from "../lib/createS3Config"
+import DynamoGateway from "../lib/DynamoGateway"
 import { isError } from "../types"
 import type { CompareLambdaEvent } from "../types/CompareLambdaEvent"
