+++ src/comparison/lambdas/compareLambda.ts
@@ -1,9 +1,9 @@
 import logger from "../../lib/logging"
-import DynamoGateway from "../DynamoGateway/DynamoGateway"
 import type { ComparisonResult } from "../lib/compareMessage"
 import compareMessage from "../lib/compareMessage"
 import createDynamoDbConfig from "../lib/createDynamoDbConfig"
 import createS3Config from "../lib/createS3Config"
+import DynamoGateway from "../lib/DynamoGateway"
 import getFileFromS3 from "../lib/getFileFromS3"
 import logInDynamoDb from "../lib/logInDynamoDb"
 import { isError } from "../types"
