+++ src/comparison/lambdas/rerunComparisonLambda.integration.test.ts
@@ -11,9 +11,9 @@ import MockDate from "mockdate"
 import MockDynamo from "tests/helpers/MockDynamo"
 import MockS3 from "tests/helpers/MockS3"
 import dynamoDbTableConfig from "tests/helpers/testDynamoDbTableConfig"
-import DynamoGateway from "../DynamoGateway/DynamoGateway"
 import createDynamoDbConfig from "../lib/createDynamoDbConfig"
 import createS3Config from "../lib/createS3Config"
+import DynamoGateway from "../lib/DynamoGateway"
 import InvokeCompareLambda from "../lib/InvokeCompareLambda"
 import type { ComparisonLog } from "../types"
 import { isError } from "../types"
