+++ src/comparison/lib/DynamoGateway.ts
