+++ src/comparison/lib/getFailedComparisonResults.ts
@@ -1,5 +1,5 @@
-import type DynamoGateway from "../DynamoGateway/DynamoGateway"
 import type { ComparisonLog, PromiseResult } from "../types"
+import type DynamoGateway from "./DynamoGateway"
 
 const getFailedComparisonResults = (dynamoGateway: DynamoGateway, limit: number): PromiseResult<ComparisonLog[]> =>
   dynamoGateway.getFailedOnes(limit)
