+++ src/comparison/lib/logInDynamoDb.integration.test.ts
@@ -3,10 +3,10 @@ import MockDate from "mockdate"
 import "tests/helpers/setEnvironmentVariables"
 import MockDynamo from "../../../tests/helpers/MockDynamo"
 import dynamoDbTableConfig from "../../../tests/helpers/testDynamoDbTableConfig"
-import DynamoGateway from "../DynamoGateway/DynamoGateway"
 import type { ComparisonLog } from "../types"
 import { isError } from "../types"
 import createDynamoDbConfig from "./createDynamoDbConfig"
+import DynamoGateway from "./DynamoGateway"
 import logInDynamoDb from "./logInDynamoDb"
 
 const config = createDynamoDbConfig()
