+++ src/comparison/lib/logInDynamoDb.ts
@@ -1,7 +1,7 @@
-import type DynamoGateway from "../DynamoGateway/DynamoGateway"
 import type { PromiseResult } from "../types"
 import { isError } from "../types"
 import type { ComparisonResult } from "./compareMessage"
+import type DynamoGateway from "./DynamoGateway"
 import getDateFromComparisonFilePath from "./getDateFromComparisonFilePath"
 
 const logInDynamoDb = async (
